+++ samples/sqltransport-sqlpersistence/Core_6/Receiver/OrderSubmittedHandler.cs
@@ -1,12 +1,9 @@
 using System.Threading.Tasks;
 using NServiceBus;
-using NServiceBus.Logging;
 
 public class OrderSubmittedHandler :
     IHandleMessages<OrderSubmitted>
 {
-    static ILog log = LogManager.GetLogger<OrderLifecycleSaga>();
-
     public Task Handle(OrderSubmitted message, IMessageHandlerContext context)
     {
         #region Reply
